Simplify PendingFileValueResolver by introducing a common trait
